+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fix BESSI content items response handling [#2820](https://github.com/rokwire/illinois-app/issues/2820).
+- Remove title from welcome video [#2831](https://github.com/rokwire/illinois-app/issues/2831).
+### Added
+- Various Notification updates [#2833](https://github.com/rokwire/illinois-app/issues/2833).
 
 ## [4.2.45] - 2022-12-20
 ### Fixed